refactors factories.py Factory classes into separate files
